refactor: AdminCommunityCreateResultPage layout파일 제거
